src/backend/requirements.txt
pandas
requests
beautifulsoup4
google-generativeai